tests: merge unit/dojson/dois in unit/dojson/hep
